pyproject.toml
[tool.poetry]
name = "pyqtfile"
version = "0.1.0"
description = ""
authors = ["Roland van Laar <roland@rolandvanlaar.nl>"]

[tool.poetry.dependencies]
python = "^3.10"
mrcrowbar = {git = "https://github.com/moralrecordings/mrcrowbar"}
numpy = "^1.25.2"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
ipython = "*"
ipdb = "*"
black = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
